% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs_wuw.R
\name{freqs_wuw}
\alias{freqs_wuw}
\title{Combine weighted freqs with unweighted ns}
\usage{
freqs_wuw(
  dataset,
  ...,
  stat = c("percent", "mean", "median", "min", "max", "quantile", "summary"),
  pr = NULL,
  nas = TRUE,
  wt = NULL,
  prompt = FALSE,
  digits = 2,
  nas_group = TRUE,
  factor_group = FALSE
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{...}{The unquoted names of a set of variables in the dataset. If nothing
is specified, the function runs a frequency on every column in given dataset.}

\item{stat}{Character, stat to run. Currently accepts 'percent,' 'mean,' 'median,' 'min,' 'max,' 'quantile,' and 'summary' (default: 'percent').}

\item{pr}{Double, for use when stat = 'quantile.' Input should be a real number x such that 0 <= x <= 100. Stands for percentile rank, which is a quantile relative to a 100-point scale. (default:NULL)}

\item{nas}{Boolean, whether or not to include NAs in the tabulation (default: TRUE).}

\item{wt}{The unquoted name of a weighting variable in the dataset (default: NULL).}

\item{prompt}{Boolean, whether or not to include the prompt in the dataset (default: FALSE).}

\item{digits}{Integer, number of significant digits for rounding (default: 2).}

\item{nas_group}{Boolean, whether or not to include NA values for the grouping variable in the tabulation (default: TRUE).}

\item{factor_group}{Boolean, whether or not to convert the grouping variable to a factor and use its labels instead of its underlying numeric values (default: FALSE)}
}
\value{
A dataframe with the variable names, prompts, values, labels, counts,
stats, and resulting calculations.
}
\description{
Use freqs_wuw() when creating a frequencies table with weighted results but unweighted ns. Saves you the time of combining the two freqs tables if run separately
}
\examples{
mtcars \%>\% y2clerk::freqs(gear)
mtcars \%>\% y2clerk::freqs(gear, wt = mpg)
mtcars \%>\% freqs_wuw(gear, wt = mpg)
}
\keyword{freqs}
