% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodology.R
\name{methodology}
\alias{methodology}
\alias{jarvis_sommelier_the_survey}
\title{Get info on survey methodology}
\usage{
methodology(
  dataset,
  population,
  start_date = StartDate,
  end_date = EndDate,
  duration = Duration__in_seconds_
)

jarvis_sommelier_the_survey(
  dataset,
  population,
  start_date = StartDate,
  end_date = EndDate,
  duration = Duration__in_seconds_
)
}
\arguments{
\item{dataset}{A dataframe for which you want the methodology information.}

\item{population}{The population size your survey is based on (e.g. number of voters in the city)}

\item{start_date}{Unquoted name of the start date column (default: StartDate)}

\item{end_date}{Unquoted name of the end date column (default: EndDate)}

\item{duration}{Unquoted name of the interview duration column in seconds (default: Duration__in_seconds_)}
}
\value{
Methodology information printed to your console
}
\description{
Get the following survey methodology information:
Length of interview (LOI)
Fielding dates
Survey margin of error (MOE)
}
\examples{
methodology(municipal_data, 30000)

}
\keyword{methodology}
