% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_names_fonts.R
\name{read_data_names_fonts}
\alias{read_data_names_fonts}
\alias{jarvis_set_the_table}
\title{prep data, names table, and fonts}
\usage{
read_data_names_fonts(
  file_name,
  fonts3 = c("flama", "Flama-Medium.otf", "Flama-Bold.otf")
)

jarvis_set_the_table(
  file_name,
  fonts3 = c("flama", "Flama-Medium.otf", "Flama-Bold.otf")
)
}
\arguments{
\item{file_name}{The file name of your data. (e.g. 'my weighted data.rds')
Using Y2 Analytics coding standards, you should have a DATA_PATH string in your environment specifying the path to your data file}

\item{fonts3}{If you want flama as your font, don't touch this argument (flama must already be installed on your computer as well).
Otherwise, it is a vector of 2-3 strings:
\enumerate{
\item Font name (e.g., 'flama')
\item File name of regular style font (e.g., 'Flama-Medium.otf')
\item File name of bold style font (e.g., 'Flama-Bold.otf')
It is also possible to specify only 2 strings. If you give only 2, the bold (what would be 3) will just be the regular (2).
DEFAULT is flama: c('flama', 'Flama-Medium.otf', 'Flama-Bold.otf')
}}
}
\value{
Two dataframes:
\enumerate{
\item responses (from your data file),
\item names (with the variable names and labels of your data)
Additionally, loads your specified font into your R environment (flama by default)
}
}
\description{
3 in 1 function for reading in data, creating a data frame of variable names/labels, and reading in custom fonts
}
\details{
Run this function at the start of your code to get your basic data/names/fonts ready
}
\examples{
\dontrun{
DATA_PATH <- '~/Dropbox (Y2 Analytics)/Y2 Analytics Team Folder/R&D/Training/orderlabel/data/'
read_data_names_fonts('training data.rds')
}
}
\keyword{font}
\keyword{names}
\keyword{read}
