% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline.R
\name{topline}
\alias{topline}
\title{Create all data for a topline}
\usage{
topline(dataset, file_name_topline, file_name_appendix, weight_var = weights)
}
\arguments{
\item{dataset}{A dataframe for which you want to create a topline}

\item{file_name_topline}{The filename you want to give your topline freqs table csv. Will be saved into your DATA_PATH}

\item{file_name_appendix}{The filename you want to give your open ends appendix csv. Will be saved into your DATA_PATH}

\item{weight_var}{DEFAULT = weights. Can be set to NULL if the data has no weights}
}
\value{
A frequencies table including all questions with the prefixes mentioned above.
}
\description{
Use topline() to automate all the frequencies for a topline report. This function can only be used if your questions have the proper prefixes:
"s_" for single select,
"m_" for multiple select,
"oe_" for open ends,
"n_" for numeric,
"r_" for ranked,
"md_" for max diff.
}
\examples{

}
\keyword{freqs}
\keyword{topline}
