% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline_appendix.R
\name{topline_appendix}
\alias{topline_appendix}
\alias{jarvis_listen_to_guests}
\title{Create open ended data for a topline}
\usage{
topline_appendix(dataset, assign_oe = NULL)

jarvis_listen_to_guests(dataset, assign_oe = NULL)
}
\arguments{
\item{dataset}{A dataframe for which you want to create a topline}

\item{assign_oe}{DEFAULT = NULL, A vector of unquoted variables to be treated as open-ended variables, put within c()}
}
\value{
A tibble of open-ended responses
}
\description{
Use topline_appendix() to automate all the qualitative responses for a topline report. This function works best if your questions have the proper prefixes and suffixes:
\enumerate{
\item prefix "oe_" for open ends
\item suffix "_TEXT" for open ends
}
}
\examples{
municipal_data \%>\%
topline_appendix()

municipal_data \%>\%
  topline_appendix(
    assign_oe = c(RecipientEmail, UserLanguage)
  )
}
\keyword{freqs}
\keyword{topline}
