% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodology.R
\name{methodology}
\alias{methodology}
\alias{jarvis_sommelier_the_survey}
\title{Get info on survey methodology}
\usage{
methodology(dataset, population)

jarvis_sommelier_the_survey(dataset, population)
}
\arguments{
\item{dataset}{A dataframe for which you want the methodology information. Your dataset needs the following standard variables from Qualtrics to work: StartDate, EndDate, Duration__in_seconds_}

\item{population}{The population size your survey is based on (e.g. number of voters in the city)}
}
\value{
Methodology information printed to your console
}
\description{
Get the following survey methodology information:
Length of interview (LOI)
Fielding dates
Survey margin of error (MOE)
}
\examples{
methodology(municipal_data, 30000)

}
\keyword{methodology}
