% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline.R
\name{topline}
\alias{topline}
\title{Create all data for a topline}
\usage{
topline(dataset, weight_var = weights)
}
\arguments{
\item{dataset}{A dataframe for which you want to create a topline}

\item{weight_var}{DEFAULT = weights. Can be set to NULL if the data has no weights}
}
\value{
2 tables saved out into your DATA_PATH folder
\enumerate{
\item A csv of frequencies for the topline called "data for topline, \emph{project name}"
\item A csv of open ends for the appendix called "data for appendix, \emph{project name}"
}
}
\description{
Use topline() to automate all the frequencies for a topline report. This function can only be used if your questions have the proper prefixes:
"s_" for single select,
"m_" for multiple select,
"oe_" for open ends,
"n_" for numeric,
"r_" for ranked,
"md_" for max diff.
You need to be using y2coding conventions and have a DATA_PATH in your R environment because topline() will save your new files to that folder
}
\examples{
\dontrun{
DATA_PATH <- '~/Desktop/'
topline(municipal_data)
}

}
\keyword{freqs}
\keyword{topline}
