% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline.R
\name{topline}
\alias{topline}
\alias{jarvis_top_us_all_off}
\title{Create all data for a topline}
\usage{
topline(dataset, weight_var)

jarvis_top_us_all_off(dataset, weight_var)
}
\arguments{
\item{dataset}{A dataframe for which you want to create a topline}

\item{weight_var}{Variable containing weights}
}
\value{
2 tables saved out into your DATA_PATH folder
\enumerate{
\item A csv of frequencies for the topline called "data for topline, \emph{project name}"
\item A csv of open ends for the appendix called "data for appendix, \emph{project name}"
}
}
\description{
Use topline() to automate all the frequencies for a topline report. This function can only be used if your questions have the proper prefixes:
\enumerate{
\item "s_" for single select,
\item "m_" for multiple select,
\item "oe_" for open ends,
\item "n_" for numeric,
\item "r_" for ranked,
\item "md_" for max diff.
You need to be using y2 coding conventions and have a DATA_PATH in your R environment because topline() will save your new files to that folder
}
}
\examples{
\dontrun{
DATA_PATH <- '~/Desktop/'
topline(municipal_data)
}

}
\keyword{freqs}
\keyword{topline}
