% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline_appendix.R
\name{topline_appendix}
\alias{topline_appendix}
\title{Create open ended data for a topline}
\usage{
topline_appendix(dataset, assign_oe = NULL, qsf)
}
\arguments{
\item{dataset}{A dataframe for which you want to create a topline}

\item{assign_oe}{DEFAULT = NULL, A vector of unquoted variables to be treated as open-ended variables, put within c()}

\item{qsf}{file path to a qsf}
}
\value{
A tibble of open-ended responses
}
\description{
Use topline_appendix() to automate all the qualitative reponses for a topline report. This function works best if your questions have the proper prefixes:
\enumerate{
\item "s_" for single select,
\item "m_" for multiple select,
\item "oe_" for open ends,
\item "n_" for numeric,
\item "r_" for ranked,
\item "md_" for max diff.
}
}
\examples{
municipal_data \%>\%
topline_appendix()

municipal_data \%>\%
  topline_appendix(
    assign_oe = c(RecipientEmail, UserLanguage)
  )
}
\keyword{freqs}
\keyword{topline}
