% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline_freqs.R
\name{topline_freqs}
\alias{topline_freqs}
\alias{jarvis_top_us_all_off}
\title{Create all data for a topline}
\usage{
topline_freqs(
  dataset,
  weight_var,
  assign_s = NULL,
  assign_m = NULL,
  assign_n = NULL,
  unweighted_ns = TRUE,
  silently = FALSE
)

jarvis_top_us_all_off(
  dataset,
  weight_var,
  assign_s = NULL,
  assign_m = NULL,
  assign_n = NULL,
  unweighted_ns = TRUE,
  silently = FALSE
)
}
\arguments{
\item{dataset}{A dataframe for which you want to create a topline}

\item{weight_var}{Variable containing weights}

\item{assign_s}{DEFAULT = NULL, A vector of unquoted variables to be treated as single select variables, put within c()}

\item{assign_m}{DEFAULT = NULL, A vector of unquoted variables to be treated as multiple select variables, put within c()}

\item{assign_n}{DEFAULT = NULL, A vector of unquoted variables to be treated as numeric variables, put within c()}

\item{unweighted_ns}{DEFAULT = TRUE, Display weighted or unweighted n-sizes in topline report}

\item{silently}{DEFAULT = FALSE, Hide message output (e.g., progress of completing freqs on variables or printing of variables not included in the topline)}
}
\value{
A tibble of frequencies
}
\description{
Use topline_freqs() to automate all the quantitative frequencies for a topline report. This function works best if your questions have the proper prefixes:
\enumerate{
\item "s_" for single select,
\item "m_" for multiple select,
\item "oe_" for open ends,
\item "n_" for numeric,
\item "r_" for ranked,
\item "md_" for max diff.
}
}
\examples{
municipal_data \%>\%
topline_freqs()

municipal_data \%>\%
  topline_freqs(
    assign_n = c(d_yearborn, Duration__in_seconds_),
    weight_var = weights
)
}
\keyword{freqs}
\keyword{topline}
